backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
psycopg==3.1.12
psycopg-binary==3.1.12
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.8.0
